Enqueue admin-exclude in new way
