+++ webpack.config.js
@@ -4,9 +4,9 @@ module.exports = {
 	...defaultConfig,
 	entry: {
 		admin: './ui/js/admin.js',
+		'admin-exclude': './ui/js/admin-exclude.js',
 		alerts: './ui/js/alerts.js',
 		'alerts-list': './ui/js/alerts-list.js',
-		exclude: './ui/js/exclude.js',
 		global: './ui/js/global.js',
 		'live-updates': './ui/js/live-updates.js',
 		settings: './ui/js/settings.js',
